mileage_app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
SQLAlchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
